Add support for indexing both during instantiation and search
Search indexes allow data that gets searched to be organised in
a more efficient manner that reduces the time taken to perform the
search.

Construction of the index generally happens before the search
is performed(as construction itself might be costly in both space
and time).

The implementation here allows indexing during instantiation
which leads to very fast searches on the same data, or search time
indexing which may or may not lead to fast searches(depends on
time it takes to construct the index).

Indexes are optional and when omitted, Seekr continues to perform
standard search algorithm on all keys / group selected keys like
before.

NOTE: Indexing only works on fields explicity specified. If the
user does not specify any field to index, Seekr never assumes and
indexes fields on it's own(although this might be a cool
optimization if during compliation some group selectors appear
multiple times and data gets indexed on it automatically once!).